Fix: remove --program-input from CLI
Problem: the program input is usually loaded through hints. These hints
depend on each program and must be reimplemented for each program. As
the CLI cannot load hints dynamically, this means that it is impossible
to load program inputs in a generic way.

Solution: remove the option entirely. If you need to run a program with
custom hints, use the Python VM `cairo-run` for now.