Feature: --bootloader-version flag (#14)
Problem: The bootloader version known by the L1 contract is different
from the one we use (0.13.0). It would be convenient to be able to
change the version of the bootloader / specify a custom one for testing.

Solution: Add a `--bootloader-version VERSION` flag. The `VERSION` value
can be one of [`0.12.3`, `0.13.0`] or be the path to a compiled
bootloader file.